Move things form calendar into their own file
