+++ lib/ui/screens/calendar_screen.dart
@@ -2,7 +2,6 @@ import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:go_router/go_router.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:intl/intl.dart';
 import 'package:reaxit/api/api_repository.dart';
@@ -10,8 +9,6 @@ import 'package:reaxit/blocs.dart';
 import 'package:reaxit/models.dart';
 import 'package:reaxit/ui/theme.dart';
 import 'package:reaxit/ui/widgets.dart';
-import 'package:sticky_headers/sticky_headers/widget.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 class CalendarScreen extends StatefulWidget {
   @override
@@ -218,91 +215,6 @@ class CalendarSearchDelegate extends SearchDelegate {
   }
 }
 
-/// _CalendarViewDay holds events attached to a day
-@visibleForTesting
-class CalendarViewDay {
-  final DateTime day;
-  final List<CalendarEvent> events;
-
-  CalendarViewDay({required this.day, required List<CalendarEvent> events})
-      : events = events.sortedBy((element) => element.start);
-}
-
-/// _CalendarViewMonth holds events attached to a month
-@visibleForTesting
-class CalendarViewMonth {
-  final DateTime month;
-  final List<CalendarViewDay> days;
-
-  CalendarViewMonth({required this.month, required List<CalendarEvent> events})
-      : days = groupBy<CalendarEvent, DateTime>(
-          events.sortedBy((element) => element.start),
-          (event) => DateTime(
-            event.start.year,
-            event.start.month,
-            event.start.day,
-          ),
-        )
-            .entries
-            .map(
-                (entry) => CalendarViewDay(day: entry.key, events: entry.value))
-            .sortedBy((element) => element.day);
-
-  List<CalendarViewDay> byDay() => days;
-}
-
-@visibleForTesting
-List<CalendarViewMonth> ensureContainsToday(
-    List<CalendarViewMonth> events, DateTime now) {
-  DateTime today = DateTime(
-    now.year,
-    now.month,
-    now.day,
-  );
-  DateTime thisMonth = DateTime(
-    now.year,
-    now.month,
-  );
-  for (var i = 0; i < events.length; i++) {
-    if (events[i].month.isAfter(thisMonth)) {
-      events.insert(i, CalendarViewMonth(month: thisMonth, events: []));
-      events[i].days.add(CalendarViewDay(day: today, events: []));
-
-      return events;
-    }
-    if (events[i].month == thisMonth) {
-      for (var j = 0; j < events[i].days.length; j++) {
-        if (events[i].days[j].day == today) return events;
-        if (events[i].days[j].day.isAfter(today)) {
-          events[i].days.insert(j, CalendarViewDay(day: today, events: []));
-          return events;
-        }
-      }
-      return events;
-    }
-  }
-  events.add(CalendarViewMonth(month: thisMonth, events: []));
-  events.last.days.add(CalendarViewDay(day: today, events: []));
-
-  return events;
-}
-
-@visibleForTesting
-List<CalendarViewMonth> groupByMonth(
-  List<CalendarEvent> eventList,
-) =>
-    groupBy<CalendarEvent, DateTime>(
-      eventList,
-      (event) => DateTime(
-        event.start.year,
-        event.start.month,
-      ),
-    )
-        .entries
-        .map(
-            (entry) => CalendarViewMonth(month: entry.key, events: entry.value))
-        .toList();
-
 /// A ScrollView that shows a calendar with [Event]s.
 ///
 /// The events are grouped by month, and date.
@@ -388,7 +300,7 @@ class CalendarScrollView extends StatelessWidget {
                 padding: const EdgeInsets.symmetric(horizontal: 12),
                 sliver: SliverList(
                   delegate: SliverChildBuilderDelegate(
-                    (_, index) => _CalendarMonth(
+                    (_, index) => CalendarMonth(
                       events: _monthGroupedEventsUp[index],
                       todayKey: todayKey,
                       thisMonthKey: thisMonthKey,
@@ -403,7 +315,7 @@ class CalendarScrollView extends StatelessWidget {
                 key: centerkey,
                 sliver: SliverList(
                   delegate: SliverChildBuilderDelegate(
-                    (_, index) => _CalendarMonth(
+                    (_, index) => CalendarMonth(
                       events: _monthGroupedEventsDown[index],
                       todayKey: todayKey,
                       thisMonthKey: thisMonthKey,
@@ -424,286 +336,6 @@ class CalendarScrollView extends StatelessWidget {
           ),
         ),
       ],
-    );
-  }
-}
-
-class _CalendarMonth extends StatelessWidget {
-  final CalendarViewMonth events;
-  final Key? todayKey;
-  final Key? thisMonthKey;
-  final DateTime now;
-
-  static final monthFormatter = DateFormat('MMMM');
-  static final monthYearFormatter = DateFormat('MMMM yyyy');
-
-  const _CalendarMonth(
-      {required this.events,
-      this.todayKey,
-      this.thisMonthKey,
-      required this.now});
-
-  @override
-  Widget build(BuildContext context) {
-    DateTime today = DateTime(
-      now.year,
-      now.month,
-      now.day,
-    );
-    DateTime thisMonth = DateTime(
-      now.year,
-      now.month,
-    );
-    return StickyHeader(
-      header: Column(
-        key: events.month == thisMonth ? thisMonthKey : null,
-        children: [
-          SizedBox(
-            width: double.infinity,
-            child: Material(
-              child: Padding(
-                padding: const EdgeInsets.symmetric(vertical: 8),
-                child: Text(
-                  events.month.year == now.year
-                      ? monthFormatter
-                          .format(events.month.toLocal())
-                          .toUpperCase()
-                      : monthYearFormatter
-                          .format(events.month.toLocal())
-                          .toUpperCase(),
-                  style: Theme.of(context).textTheme.titleMedium,
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(height: 8),
-        ],
-      ),
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          for (final day in events.byDay())
-            _DayCard(
-                day: day.day,
-                events: day.events,
-                now: now,
-                key: day.day == today ? todayKey : null),
-        ],
-      ),
-    );
-  }
-}
-
-class _DayCard extends StatelessWidget {
-  final DateTime day;
-  final DateTime now;
-  final List<Widget> eventWidgets;
-
-  static final dayFormatter = DateFormat(DateFormat.ABBR_WEEKDAY);
-
-  _DayCard(
-      {required DateTime day,
-      required List<CalendarEvent> events,
-      required this.now,
-      Key? key})
-      : eventWidgets = events.map((event) => _EventCard(event)).toList(),
-        day = day.toLocal(),
-        super(key: key ?? ValueKey(day));
-
-  @override
-  Widget build(BuildContext context) {
-    DateTime today = DateTime(
-      now.year,
-      now.month,
-      now.day,
-    );
-
-    final textTheme = Theme.of(context).textTheme;
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        SizedBox(
-          width: 60,
-          child: Padding(
-            padding: const EdgeInsets.only(right: 12, top: 4),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.start,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: [
-                Text(
-                  dayFormatter.format(day).toUpperCase(),
-                  style: textTheme.bodySmall!.apply(
-                      color: day == today
-                          ? magenta
-                          : textTheme.bodySmall!.color!.withOpacity(0.5)),
-                ),
-                Text(
-                  day.day.toString(),
-                  style: textTheme.displaySmall!.apply(
-                      color: day == today
-                          ? magenta
-                          : textTheme.displaySmall!.color!.withOpacity(0.5)),
-                  strutStyle: const StrutStyle(
-                    forceStrutHeight: true,
-                    leading: 2.2,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-        Expanded(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: eventWidgets.isNotEmpty
-                ? eventWidgets
-                : [
-                    Center(
-                      child: Text(
-                        'There are no events this day',
-                        style: TextStyle(
-                          color:
-                              textTheme.displaySmall!.color!.withOpacity(0.5),
-                          fontWeight: FontWeight.bold,
-                          fontSize: 20,
-                        ),
-                        strutStyle: const StrutStyle(
-                          forceStrutHeight: true,
-                          leading: 4,
-                        ),
-                      ),
-                    )
-                  ],
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _EventCard extends StatelessWidget {
-  final CalendarEvent event;
-
-  _EventCard(this.event) : super(key: ObjectKey(event));
-
-  void openEvent(BuildContext context) {
-    if (event.parentEvent is PartnerEvent) {
-      launchUrl(
-        (event.parentEvent as PartnerEvent).url,
-        mode: LaunchMode.externalApplication,
-      );
-    } else {
-      context.pushNamed(
-        'event',
-        params: {'eventPk': event.pk.toString()},
-        extra: event.parentEvent,
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    Color color;
-    if (event.parentEvent is Event &&
-        (event.parentEvent as Event).isRegistered) {
-      color = Theme.of(context).highlightColor;
-    } else if (event.parentEvent is PartnerEvent) {
-      color = Colors.black;
-    } else {
-      color = Colors.grey[800]!;
-    }
-
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 12),
-      child: Material(
-        borderRadius: const BorderRadius.all(Radius.circular(2)),
-        type: MaterialType.card,
-        color: color,
-        child: InkWell(
-          onTap: () => openEvent(context),
-          // Prevent painting ink outside of the card.
-          borderRadius: const BorderRadius.all(Radius.circular(2)),
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              mainAxisAlignment: MainAxisAlignment.start,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  event.title,
-                  maxLines: 2,
-                  overflow: TextOverflow.ellipsis,
-                  style: const TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                Text(
-                  event.label,
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                  style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                        color: Colors.white.withOpacity(0.8),
-                      ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class AnimatedLoader extends StatefulWidget {
-  final bool visible;
-  const AnimatedLoader({super.key, required this.visible});
-
-  @override
-  State<AnimatedLoader> createState() => _AnimatedLoaderState();
-}
-
-/// AnimationControllers can be created with `vsync: this` because of TickerProviderStateMixin.
-class _AnimatedLoaderState extends State<AnimatedLoader>
-    with TickerProviderStateMixin {
-  late final AnimationController _controller = AnimationController(
-    duration: const Duration(milliseconds: 200),
-    vsync: this,
-  );
-  late final Animation<double> _animation = CurvedAnimation(
-    parent: _controller,
-    curve: Curves.fastOutSlowIn,
-  );
-
-  @override
-  void didUpdateWidget(AnimatedLoader oldWidget) {
-    if (widget.visible) {
-      _controller.value = 1;
-    } else {
-      _controller.reverse();
-    }
-    super.didUpdateWidget(oldWidget);
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SizeTransition(
-      sizeFactor: _animation,
-      axis: Axis.vertical,
-      child: const Center(
-        child: Padding(
-            padding: EdgeInsets.all(12), child: CircularProgressIndicator()),
-      ),
     );
   }
 }
